Refactor: Reorganize pipeline subpackages
